Move helpers functions to 'internal' and update code to use them
