+++ internal/pointer/pointer_test.go
@@ -1,10 +1,9 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package v1alpha2
+package pointer
 
 import (
-	"reflect"
 	"testing"
 )
 
@@ -25,22 +24,5 @@ func TestPointerOf(t *testing.T) {
 	i64p := PointerOf(i64)
 	if i64 != *i64p {
 		t.Error("Failed to get int64 pointer")
-	}
-}
-
-func TestRemoveFromSlice(t *testing.T) {
-	input := [][]any{
-		{1, 2, 3},
-		{"a", "b", "c"},
-	}
-	want := [][]any{
-		{1, 3},
-		{"a", "c"},
-	}
-	for i, s := range input {
-		r := RemoveFromSlice(s, 1)
-		if !reflect.DeepEqual(want[i], r) {
-			t.Errorf("Failed to remove an element from slice %d", i)
-		}
 	}
 }
